Protect against disabling the last admin account
This adds extra checks in to ensure that a user cannot
disable themselves if they are the last admin account.

Combined with the existing check against demoting the last admin and automatically skipping yourself when bulk deleting this should prevent a user from locking themselves out of their own admin.