Separate the broadcast URLs into 2 different URLs
We need both a public and internal URL for the broadcast system.

Public URL:
Serves the client.js file for Faye and will be accessible on the internet.

In development mode this will likely be http://localhost:4242/stream

In production it will likely be your domain name instead along with a path setup in nginx to distinguish it.

Internal URL:
This is the address that we'll POST to from the app.

In development mode it will likely be http://localhost.com:4242/stream.

In production mode it likely be linked to the Faye server over DNS, an IP address or through a Docker link.