Add robots.txt and move page icons to the public dir
Certain static assets such as the page's favicon were moved to the public directory. This is where assets will get built to in non-development settings.

A proper web server will be able to serve them directly now. It also sets the stage for where our custom error templates will be stored in the future.