Use Alembic directly for database migrations
The goal of this project is to save you time. Alembic has a lot of great documentation and features.

It was too tedious to try and wrap everything that Alembic does. It also allows for you to follow along with Alembic tutorials.

The only "cost" is it's no longer unified into the CLI, but I think that trade off is worth it.