Change Dockerfile to use slim instead of alpine
Alpine was causing show stopping file syncing
issues on OSX with the Docker Toolbox (1.11.x)

If you tried to mount in a volume to let's say
update app code without re-building images,
then the Docker Machine would not see the sync
results.

It also caused null bytes to be written to files
which in turn caused applications to not work.